Move fire to _vendored submodule, reexport
